fix(entity): cast (short) 0 dans le builder OtpCode pour corriger l'erreur de compilation

+++ backend/src/main/java/cm/imf/pipeline/service/AuthService.java
@@ -134,7 +134,7 @@ public class AuthService implements IAuthService {
                 .user(user)
                 .codeHash(hash(code))
                 .expiresAt(OffsetDateTime.now().plusMinutes(10))
-                .attemptsUsed(0)
+                .attemptsUsed((short) 0)
                 .used(false)
                 .build());
